+++ CHANGELOG.md
@@ -268,7 +268,7 @@ FEATURES
 FEATURES
 
 * Updated to support AWX version `21.9.0`
-* Added manual peering scenario (Ansible Automation Platform style)
+* Added manual peering scenario (Automation Platform style)
 * Added hop node support
 
 IMPROVEMENTS